feat(crypto): prekeys live in the key vault; blob v2 authenticates the key type (F-1b phase 2)
Signed prekeys, one-time prekeys and the ML-KEM-768 signed prekey are now
generated and used inside the key vault on mobile and desktop. JS only holds
the sealed stored form ("vault1:" + blob) and short-lived handles.

- Blob v2: the key type is inside the secretbox and must match the header
  type, so a relabelled blob (e.g. identity key claiming to be a prekey) is
  rejected. v1 blobs are refused.
- New key type x25519prekey: does every X25519 operation of an x25519 key,
  but can be exported with purpose "deviceSync" (SPK to our own linked
  devices) without the desktop confirmation dialog. Identity keys still need
  the native dialog; deriveEd25519 requires an exact x25519 key.
- X3DH receiver and Double Ratchet take SecretRef (bytes or vault handle);
  ratchetSerde refuses to persist a state that still holds a handle.
- Desktop ML-KEM moved from the renderer to the main process; a guard test
  forbids @noble/post-quantum in the renderer.
- Existing raw prekey rows are migrated to the stored form on load.
- Vault handle cache invalidated by vault.epoch() on lock/destroy.

Tests: C differential (export, copy, relabel, v1 reject, prekey) incl. ASan;
mobile vault/x3dh/ensureDevicePreKeys/ratchetSerde; desktop vault ops export
gate, deviceSync, relabel, crypto-imports guard.

+++ mobile/src/socket/__tests__/client.desyncRecovery.test.ts
@@ -191,7 +191,7 @@ function buildDesyncedEnvelope(me: Identity, peer: Identity, createdAtMs: number
   desyncedState.createdAtMs = createdAtMs;
   const serial = {
     RK: Array.from(desyncedState.RK),
-    DHs: { publicKey: Array.from(desyncedState.DHs.publicKey), secretKey: Array.from(desyncedState.DHs.secretKey) },
+    DHs: { publicKey: Array.from(desyncedState.DHs.publicKey), secretKey: Array.from(desyncedState.DHs.secretKey as Uint8Array) },
     DHr: desyncedState.DHr ? Array.from(desyncedState.DHr) : null,
     CKs: desyncedState.CKs ? Array.from(desyncedState.CKs) : null,
     CKr: desyncedState.CKr ? Array.from(desyncedState.CKr) : null,
